Fixed interval getting jest stuck

+++ game_service/game.js
@@ -33,12 +33,14 @@ mongodb();
 // Save questions for each 24 hours
 loadInitialQuestions();
 
-setInterval( async () =>
-{
-  await deleteOlderQuestions();
-  await saveQuestionsInDB();
-}, 24 * 60 * 60 * 1000);
-
+//We dont want to do this in a test enviroment
+if (process.env.DB_URL) {
+  setInterval( async () =>
+  {
+    await deleteOlderQuestions();
+    await saveQuestionsInDB();
+  }, 24 * 60 * 60 * 1000);
+}
 // Start the server
 const server = app.listen(port, () => {
   sync();
